SRM 255 WordCompositionGame
Three players are playing a game of word composition in which each
player writes down a list of words. After the time expires the scores
are calculated using the following rules. Each player gains 3 points for
each unique word that only he has, 2 points for each word that is shared
with exactly one other player, and 1 point for each word that is shared
with both of the other players.

You will be given String[]s listA, listB and listC - the word lists of
playerA, playerB and playerC respectively. Your method should return
scores of playerA, playerB and playerC in the form
"scoreA/scoreB/scoreC" (quotes for clarity).