SRM 518 LargestSubsequence
For Strings x and y, we say y is a subsequence of x if y can be obtained
from x by erasing some (possibly all or none) of the letters in x. For
example, "tpcdr" is a subsequence of "topcoder", while "rt" is not.

Given a String s, return the lexicographically largest subsequence of s.