SRM 291 Snowflakes
Paper snowflakes make great decorations and they are easy to make. Take
a sheet of square checkered paper with an even number of squares on each
side. Fold it in half, folding the upper part up and over to the bottom
part, so that you have a rectangle. Now fold that rectangle in half
lengthwise, from left to right, to make a smaller square. Next, take the
right upper corner of that square and fold it up and over to the left
bottom corner, forming a triangle (see pictures below, cells which stay
immovable after all foldings are marked red on the fifth picture). Cut
out some squares along grid lines (as shown on the last picture).