SRM 422 PrimeSoccer
You are watching a soccer match, and you wonder what the probability is
that at least one of the two teams will score a prime number of goals.
The game lasts 90 minutes, and to simplify the analysis, we will split
the match into five-minute intervals. The first interval is the first
five minutes, the second interval is the next five minutes, and so on.
During each interval, there is a skillOfTeamA percent probability that
team A will score a goal, and a skillOfTeamB percent probability that
teamB will score a goal. Assume that each team will score at most one
goal within each interval. Return the probability that at least one team
will have a prime number as its final score.