SRM 384 Library
Mr. Agent has entered a government library where secret documents are
stored. Documents are stored in different rooms, and each document has a
list of user groups that can access it. In order to access a document, a
person must have access to the room in which the document is stored and
also be a member of at least one user group with access to the document.
Your task is to determine the number of documents Mr. Agent can access.